agent/pyproject.toml
[tool.poetry]
name = "agent"
version = "0.1.0"
description = ""
authors = ["nag763 <loic.labeye@tutanota.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
google-adk = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
